Define `appsignal.start()` initialization method
This allows for a more ergonomic replacement to the previous:

```python
from __appsignal__ import appsignal
appsignal.start()
```

Into a simpler:

```python
import appsignal
appsignal.start()
```

But, more importantly, it allows users to keep the AppSignal
configuration file at the project root (where commands such as
`appsignal diagnose` expect to find it) and refer it from other
parts of their application (such as `project_name/wsgi.py` in a
Django project) without imposing the demand for the project root
to itself be a Python package.
